some minor refactoring of tracker records structs
This makes consistent use of `mut &[u8]` as the incomming type for
parsing these records to ensure that this can be used to generically
read from any u8 byte array and not require the caller to use
`BytesMut`. Hopefully this works well.

With that change, we no longer properly advance the cursor, though, so
we added a `data_size()` function on the struct which returns the number
of bytes that that struct requires. This can be used after parsing to
know how many bytes of data we read so the cursor can be advanced.

also added `put_slice()` functions for writing into a buffer.