new hotline_tracker::Header struct
this is the header used in the tracker protocol and previously, this
logic was hard-coded in different ways where it was used because it
doesn't really change.

but in the interest of code maintainability and flexibility, this has
been broken out into a module. since we're not dealing with hundreds of
requests per second, the overhead should be minimal and we could always
re-optimise this later.