reworked the config code again - default to current directory
previously, we'd kinda go around in a circle, trying to find a location
and then default on the user's .config directory. but this isn't really
where we want to automatically create the database file. Instead, we
want to do it in the current directory.

Also, the config itself shouldn't care how it's being called, so it
no longer accepts an argument (which we were passing the CLI argument
value). We let the CLI part of the code deal with setting the config
path and totally side-step that function.

Additionally, a DEFAULT_CONFIG_FILENAME constant was added so we don't
repeat the `tracker.toml` string everywhere.